Move Node and NodeTree to their own file.
NodeTree was formerly but shortly called TreeBase
